+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Arm64 and FreeBSD support for several roles.
+- Doas super command support.
 - Group install flag.
 - Tools for Libvirt role.
 - Yazi role.
@@ -17,7 +18,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Renamed Sudo role to Super.
-- Unix shell scripts to be POSIX compatible.
+- Most unix shell scripts to be POSIX compatible.
 
 ### Fixed